Renamed files & Updated save() in train.py
