+++ data_generator.py
@@ -1,15 +1,13 @@
+import keras
 import numpy as np
 import nibabel as nib
-import keras
+from pathlib import Path
 
-from data_loader import *
 
-import numpy as np
-import keras
 
 class DataGenerator(keras.utils.Sequence):
     'Generates data for Keras'
-    def __init__(self, examples, labels, batch_size=1, dim=(256,256), n_channels=1,
+    def __init__(self, examples, labels, batch_size=1, dim=(256,256,256), n_channels=1,
                  n_classes=1, shuffle=True):
         'Initialization'
         self.dim = dim
@@ -18,8 +16,13 @@ class DataGenerator(keras.utils.Sequence):
         self.list_IDs = examples
         self.n_channels = n_channels
         self.shuffle = shuffle
+        # Add an index var. to keep track of location
+        self.index = 0
         self.on_epoch_end()
 
+    def next(self):
+        'Get the next element in the generator'
+        return self.__getitem__(self.index)
 
     def __len__(self):
         'Denotes the number of batches per epoch'
@@ -36,7 +39,8 @@ class DataGenerator(keras.utils.Sequence):
         
         # Generate data
         X, y = self.__data_generation(list_IDs_temp, list_ys_temp)
-
+        # Update index
+        self.index+=1
         return X, y
 
     def on_epoch_end(self):
@@ -51,36 +55,52 @@ class DataGenerator(keras.utils.Sequence):
         min_val = np.min(x)
         norm_x = (x-min_val)/(max_val - min_val + 1e-7)
         return norm_x
-    
+
     def padImage(self, img, mask=False):
-        x_dim = img.shape[0]
-        y_dim = img.shape[1]
-        if x_dim == y_dim:
-            return img
-        else:
-            pad_w = x_dim - y_dim
-            if mask:
-                return np.pad(img, [(0,0), (pad_w//2,pad_w//2), (0,0)],
-                              mode='constant',
-                              constant_values=(1))
-            else:
-                return np.pad(img, [(0,0), (pad_w//2,pad_w//2), (0,0)],
-                              mode='constant')
+        '''
+        Input -> img : numpy array, mask : Boolean
+        ''' 
+        img_c = img.copy()
+        # Get max dim
+        dims = img.shape
+        max_dim = max(dims)
+        # Padding Args.
+        pad_value = 0
+        pad_arr = [(0,0), (0,0), (0,0)]
+        # Check if mask or input to pad w. 1 or 0
+        if mask:
+            pad_value = 1
+        # Check if other dims eq. largest dim
+        for idx, dim in enumerate(dims):
+            if dim != max_dim:
+                pad_w = max_dim - dim
+                pad_arr[idx] = (pad_w//2, pad_w//2)
+                img_c = np.pad(img, pad_arr,
+                                mode='constant', constant_values=(pad_value))
+
+        return img_c
+
                 
                 
 
     def __data_generation(self, list_IDs_temp, list_ys_temp):
-        'Generates data containing batch_size samples' # X : (n_samples, *dim, n_channels)
+        'Generates data containing batch_size samples'
         # Initialization
-        X = np.empty((self.batch_size, *self.dim, self.n_channels))
-        y = np.empty((self.batch_size, *self.dim, self.n_channels))
+        X = np.empty((self.batch_size, *self.dim))
+        y = np.empty((self.batch_size, *self.dim))
 
         # Generate data
         for i, ID in enumerate(list_IDs_temp):
+            # @TODO remove when file name consistent
+            ID = ID.replace('.gz', '')
+            # Check if file exists, continue if not
+            if not Path(ID).is_file():
+                print(ID, ' is missing.') 
+                continue
             # Store sample
-            X[i, ] = self.padImage(self.normalizeImg((nib.load(ID).get_data().astype(np.float32))[:, :, np.newaxis]))
+            X[i, ] = self.padImage(self.normalizeImg(nib.load(ID).get_data().astype(np.float32)))
             # Store sample segmentation
-            y[i, ]  = self.padImage((nib.load(list_ys_temp[i]).get_data().astype(np.float32))[:, :, np.newaxis], mask=True)
+            y[i, ]  = self.padImage((nib.load(list_ys_temp[i]).get_data().astype(np.float32)), mask=True)
 
         return X, y
