+++ unet.py
@@ -11,12 +11,7 @@ from keras.callbacks import ModelCheckpoint, LearningRateScheduler, Callback, CS
 from keras import backend as keras
 from keras.utils.training_utils import multi_gpu_model
 
-
-from data_loader import *
-from data_generator import DataGenerator
 from metrics import dice_coef
-from losses import dice_coef_loss
-
         
 class myUnet(object):
 
@@ -24,155 +19,21 @@ class myUnet(object):
         self.img_rows = img_rows
         self.img_cols = img_cols
 
-    def load_data(self):
-        pass
-    
-    def get_unet_basic2(self):
-        
-        inputs = Input((self.img_rows, self.img_cols, 1))
-        conv1 = Conv2D(32, (3, 3), activation='relu', padding='same')(inputs)
-        conv1 = Conv2D(32, (3, 3), activation='relu', padding='same')(conv1)
-        pool1 = MaxPooling2D(pool_size=(2, 2))(conv1)
-
-        conv2 = Conv2D(64, (3, 3), activation='relu', padding='same')(pool1)
-        conv2 = Conv2D(64, (3, 3), activation='relu', padding='same')(conv2)
-        pool2 = MaxPooling2D(pool_size=(2, 2))(conv2)
-
-        conv3 = Conv2D(128, (3, 3), activation='relu', padding='same')(pool2)
-        conv3 = Conv2D(128, (3, 3), activation='relu', padding='same')(conv3)
-        pool3 = MaxPooling2D(pool_size=(2, 2))(conv3)
-
-        conv4 = Conv2D(256, (3, 3), activation='relu', padding='same')(pool3)
-        conv4 = Conv2D(256, (3, 3), activation='relu', padding='same')(conv4)
-        pool4 = MaxPooling2D(pool_size=(2, 2))(conv4)
-
-        conv5 = Conv2D(512, (3, 3), activation='relu', padding='same')(pool4)
-        conv5 = Conv2D(512, (3, 3), activation='relu', padding='same')(conv5)
-
-        up6 = concatenate([Conv2DTranspose(256, (2, 2), strides=(2, 2), padding='same')(conv5), conv4], axis=3)
-        conv6 = Conv2D(256, (3, 3), activation='relu', padding='same')(up6)
-        conv6 = Conv2D(256, (3, 3), activation='relu', padding='same')(conv6)
-
-        up7 = concatenate([Conv2DTranspose(128, (2, 2), strides=(2, 2), padding='same')(conv6), conv3], axis=3)
-        conv7 = Conv2D(128, (3, 3), activation='relu', padding='same')(up7)
-        conv7 = Conv2D(128, (3, 3), activation='relu', padding='same')(conv7)
-
-        up8 = concatenate([Conv2DTranspose(64, (2, 2), strides=(2, 2), padding='same')(conv7), conv2], axis=3)
-        conv8 = Conv2D(64, (3, 3), activation='relu', padding='same')(up8)
-        conv8 = Conv2D(64, (3, 3), activation='relu', padding='same')(conv8)
-
-        up9 = concatenate([Conv2DTranspose(32, (2, 2), strides=(2, 2), padding='same')(conv8), conv1], axis=3)
-        conv9 = Conv2D(32, (3, 3), activation='relu', padding='same')(up9)
-        conv9 = Conv2D(32, (3, 3), activation='relu', padding='same')(conv9)
-
-        conv10 = Conv2D(1, (1, 1), activation='sigmoid')(conv9)
-
-        model = Model(inputs=[inputs], outputs=[conv10])
-
-        model.compile(optimizer=Adam(lr=1e-5), loss='binary_crossentropy', metrics=['accuracy', 'mse'])
-
-        return model
-    
-    
-    def get_unet_basic(self):
-        '''unet with crop(because padding = valid)'''
-        
-        inputs = Input((self.img_rows, self.img_cols,1))
-
-        conv1 = Conv2D(64, 3, activation = 'relu', padding = 'valid', kernel_initializer = 'he_normal')(inputs)
-        print("conv1 shape:",conv1.shape)
-        conv1 = Conv2D(64, 3, activation = 'relu', padding = 'valid', kernel_initializer = 'he_normal')(conv1)
-        print("conv1 shape:",conv1.shape)
-        crop1 = Cropping2D(cropping=((90,90),(90,90)))(conv1)
-        print("crop1 shape:",crop1.shape)
-        pool1 = MaxPooling2D(pool_size=(2, 2))(conv1)
-        print("pool1 shape:",pool1.shape)
-
-        conv2 = Conv2D(128, 3, activation = 'relu', padding = 'valid', kernel_initializer = 'he_normal')(pool1)
-        print("conv2 shape:",conv2.shape)
-        conv2 = Conv2D(128, 3, activation = 'relu', padding = 'valid', kernel_initializer = 'he_normal')(conv2)
-        print("conv2 shape:",conv2.shape)
-        crop2 = Cropping2D(cropping=((41,41),(41,41)))(conv2)
-        print("crop2 shape:",crop2.shape)
-        pool2 = MaxPooling2D(pool_size=(2, 2))(conv2)
-        print("pool2 shape:",pool2.shape)
-
-        conv3 = Conv2D(256, 3, activation = 'relu', padding = 'valid', kernel_initializer = 'he_normal')(pool2)
-        print("conv3 shape:",conv3.shape)
-        conv3 = Conv2D(256, 3, activation = 'relu', padding = 'valid', kernel_initializer = 'he_normal')(conv3)
-        print("conv3 shape:",conv3.shape)
-        crop3 = Cropping2D(cropping=((16,17),(16,17)))(conv3)
-        print("crop3 shape:",crop3.shape)
-        pool3 = MaxPooling2D(pool_size=(2, 2))(conv3)
-        print("pool3 shape:",pool3.shape)
-
-        conv4 = Conv2D(512, 3, activation = 'relu', padding = 'valid', kernel_initializer = 'he_normal')(pool3)
-        conv4 = Conv2D(512, 3, activation = 'relu', padding = 'valid', kernel_initializer = 'he_normal')(conv4)
-        print("conv4 shape:", conv4.shape)
-        #drop4 = Dropout(0.5)(conv4)
-        crop4 = Cropping2D(cropping=((4,4),(4,4)))(conv4)
-        print('crop4 shape:', crop4.shape)
-        pool4 = MaxPooling2D(pool_size=(2, 2))(conv4)
-
-        conv5 = Conv2D(1024, 3, activation = 'relu', padding = 'valid', kernel_initializer = 'he_normal')(pool4)
-        conv5 = Conv2D(1024, 3, activation = 'relu', padding = 'valid', kernel_initializer = 'he_normal')(conv5)
-        #drop5 = Dropout(0.5)(conv5)
-
-        up6 = Conv2D(512, 2, activation = 'relu', padding = 'same', kernel_initializer = 'he_normal')(UpSampling2D(size = (2,2))(conv5))
-        merge6 = merge([crop4,up6], mode = 'concat', concat_axis = 3)
-        conv6 = Conv2D(512, 3, activation = 'relu', padding = 'valid', kernel_initializer = 'he_normal')(merge6)
-        conv6 = Conv2D(512, 3, activation = 'relu', padding = 'valid', kernel_initializer = 'he_normal')(conv6)
-
-        up7 = Conv2D(256, 2, activation = 'relu', padding = 'same', kernel_initializer = 'he_normal')(UpSampling2D(size = (2,2))(conv6))
-        merge7 = merge([crop3,up7], mode = 'concat', concat_axis = 3)
-        conv7 = Conv2D(256, 3, activation = 'relu', padding = 'valid', kernel_initializer = 'he_normal')(merge7)
-        conv7 = Conv2D(256, 3, activation = 'relu', padding = 'valid', kernel_initializer = 'he_normal')(conv7)
-
-        up8 = Conv2D(128, 2, activation = 'relu', padding = 'same', kernel_initializer = 'he_normal')(UpSampling2D(size = (2,2))(conv7))
-        merge8 = merge([crop2,up8], mode = 'concat', concat_axis = 3)
-        conv8 = Conv2D(128, 3, activation = 'relu', padding = 'valid', kernel_initializer = 'he_normal')(merge8)
-        conv8 = Conv2D(128, 3, activation = 'relu', padding = 'valid', kernel_initializer = 'he_normal')(conv8)
-
-        up9 = Conv2D(64, 2, activation = 'relu', padding = 'same', kernel_initializer = 'he_normal')(UpSampling2D(size = (2,2))(conv8))
-        print("up9 shape:", up9.shape)
-        merge9 = merge([crop1,up9], mode = 'concat', concat_axis = 3)
-        conv9 = Conv2D(64, 3, activation = 'relu', padding = 'valid', kernel_initializer = 'he_normal')(merge9)
-        conv9 = Conv2D(64, 3, activation = 'relu', padding = 'valid', kernel_initializer = 'he_normal')(conv9)
-        conv9 = Conv2D(2, 3, activation = 'relu', padding = 'valid', kernel_initializer = 'he_normal')(conv9)
-        # Convert to sigmoid output
-        conv10 = Conv2D(1, 1, activation = 'sigmoid')(conv9)
-
-        model = Model(input = inputs, output = conv10)
-
-        model.compile(optimizer = Adam(lr = 1e-4), loss = 'binary_crossentropy', metrics = ['accuracy','mse'])
-        
-        return model
-        
-
-    def get_unet_zhi(self):
+    def get_unet(self):
 
         inputs = Input((self.img_rows, self.img_cols,1))
 
         conv1 = Conv2D(64, 3, activation = 'relu', padding = 'same', kernel_initializer = 'he_normal')(inputs)
-        #print "conv1 shape:",conv1.shape
         conv1 = Conv2D(64, 3, activation = 'relu', padding = 'same', kernel_initializer = 'he_normal')(conv1)
-        #print "conv1 shape:",conv1.shape
         pool1 = MaxPooling2D(pool_size=(2, 2))(conv1)
-        #print "pool1 shape:",pool1.shape
 
         conv2 = Conv2D(128, 3, activation = 'relu', padding = 'same', kernel_initializer = 'he_normal')(pool1)
-        #print "conv2 shape:",conv2.shape
         conv2 = Conv2D(128, 3, activation = 'relu', padding = 'same', kernel_initializer = 'he_normal')(conv2)
-        #print "conv2 shape:",conv2.shape
         pool2 = MaxPooling2D(pool_size=(2, 2))(conv2)
-        #print "pool2 shape:",pool2.shape
 
         conv3 = Conv2D(256, 3, activation = 'relu', padding = 'same', kernel_initializer = 'he_normal')(pool2)
-        #print "conv3 shape:",conv3.shape
         conv3 = Conv2D(256, 3, activation = 'relu', padding = 'same', kernel_initializer = 'he_normal')(conv3)
-        #print "conv3 shape:",conv3.shape
         pool3 = MaxPooling2D(pool_size=(2, 2))(conv3)
-        #print "pool3 shape:",pool3.shape
 
         conv4 = Conv2D(512, 3, activation = 'relu', padding = 'same', kernel_initializer = 'he_normal')(pool3)
         conv4 = Conv2D(512, 3, activation = 'relu', padding = 'same', kernel_initializer = 'he_normal')(conv4)
@@ -206,8 +67,8 @@ class myUnet(object):
         conv10 = Conv2D(1, 1, activation = 'sigmoid')(conv9)
 
         model = Model(input = inputs, output = conv10)
-        #multi_model = multi_gpu_model(model, gpus=2)
-        model.compile(optimizer = Adam(lr = 1e-4), loss = 'binary_crossentropy', metrics = ['accuracy','mse'])
+        #@TODO add dice coefficient as custom metric
+        model.compile(optimizer = Adam(lr = 1e-4), loss = 'binary_crossentropy', metrics = ['accuracy'])#dice_coef])
         
         return model
     
@@ -225,8 +86,6 @@ class myUnet(object):
                                   view=view,
                                   set_type=set_type)
             
-            
-    
     def save_predictions(self, file_names, predictions, view, set_type):
         # Iterate through filenames and save predictions
         for idx, fn in enumerate(file_names):
@@ -235,18 +94,14 @@ class myUnet(object):
             #print('predicted fn : ', pred_fp)
             self.save_img(predictions[idx], fn = pred_fp)
             print(pred_fp, ' saved!') 
-            
 
     def save_img(self, img, fp = '.', fn = 'd_mask.nii'):
         img_nii = nib.Nifti1Image(img, np.eye(4))
         img_nii.to_filename(os.path.join(fp, fn))
         return True
 
-
-
-
 if __name__ == '__main__':
     myunet = myUnet()
-    myunet.predict_side()
+    myunet.get_unet()
