--- unet_refactored.py
@@ -1,107 +1,0 @@
-# Force Keras to use CPU
-import os
-# Enable to run on cpu only
-#os.environ["CUDA_DEVICE_ORDER"] = "PCI_BUS_ID"
-#os.environ["CUDA_VISIBLE_DEVICES"] = ""
-import numpy as np
-from keras.models import *
-from keras.layers import Input, merge, Conv2D, MaxPooling2D, UpSampling2D, Dropout, Cropping2D, concatenate, Conv2DTranspose
-from keras.optimizers import *
-from keras.callbacks import ModelCheckpoint, LearningRateScheduler, Callback, CSVLogger
-from keras import backend as keras
-from keras.utils.training_utils import multi_gpu_model
-
-from metrics import dice_coef
-        
-class myUnet(object):
-
-    def __init__(self, img_rows = 256, img_cols = 256):
-        self.img_rows = img_rows
-        self.img_cols = img_cols
-
-    def get_unet(self):
-
-        inputs = Input((self.img_rows, self.img_cols,1))
-
-        conv1 = Conv2D(64, 3, activation = 'relu', padding = 'same', kernel_initializer = 'he_normal')(inputs)
-        conv1 = Conv2D(64, 3, activation = 'relu', padding = 'same', kernel_initializer = 'he_normal')(conv1)
-        pool1 = MaxPooling2D(pool_size=(2, 2))(conv1)
-
-        conv2 = Conv2D(128, 3, activation = 'relu', padding = 'same', kernel_initializer = 'he_normal')(pool1)
-        conv2 = Conv2D(128, 3, activation = 'relu', padding = 'same', kernel_initializer = 'he_normal')(conv2)
-        pool2 = MaxPooling2D(pool_size=(2, 2))(conv2)
-
-        conv3 = Conv2D(256, 3, activation = 'relu', padding = 'same', kernel_initializer = 'he_normal')(pool2)
-        conv3 = Conv2D(256, 3, activation = 'relu', padding = 'same', kernel_initializer = 'he_normal')(conv3)
-        pool3 = MaxPooling2D(pool_size=(2, 2))(conv3)
-
-        conv4 = Conv2D(512, 3, activation = 'relu', padding = 'same', kernel_initializer = 'he_normal')(pool3)
-        conv4 = Conv2D(512, 3, activation = 'relu', padding = 'same', kernel_initializer = 'he_normal')(conv4)
-        drop4 = Dropout(0.5)(conv4)
-        pool4 = MaxPooling2D(pool_size=(2, 2))(drop4)
-
-        conv5 = Conv2D(1024, 3, activation = 'relu', padding = 'same', kernel_initializer = 'he_normal')(pool4)
-        conv5 = Conv2D(1024, 3, activation = 'relu', padding = 'same', kernel_initializer = 'he_normal')(conv5)
-        drop5 = Dropout(0.5)(conv5)
-
-        up6 = Conv2D(512, 2, activation = 'relu', padding = 'same', kernel_initializer = 'he_normal')(UpSampling2D(size = (2,2))(drop5))
-        merge6 = merge([drop4,up6], mode = 'concat', concat_axis = 3)
-        conv6 = Conv2D(512, 3, activation = 'relu', padding = 'same', kernel_initializer = 'he_normal')(merge6)
-        conv6 = Conv2D(512, 3, activation = 'relu', padding = 'same', kernel_initializer = 'he_normal')(conv6)
-
-        up7 = Conv2D(256, 2, activation = 'relu', padding = 'same', kernel_initializer = 'he_normal')(UpSampling2D(size = (2,2))(conv6))
-        merge7 = merge([conv3,up7], mode = 'concat', concat_axis = 3)
-        conv7 = Conv2D(256, 3, activation = 'relu', padding = 'same', kernel_initializer = 'he_normal')(merge7)
-        conv7 = Conv2D(256, 3, activation = 'relu', padding = 'same', kernel_initializer = 'he_normal')(conv7)
-
-        up8 = Conv2D(128, 2, activation = 'relu', padding = 'same', kernel_initializer = 'he_normal')(UpSampling2D(size = (2,2))(conv7))
-        merge8 = merge([conv2,up8], mode = 'concat', concat_axis = 3)
-        conv8 = Conv2D(128, 3, activation = 'relu', padding = 'same', kernel_initializer = 'he_normal')(merge8)
-        conv8 = Conv2D(128, 3, activation = 'relu', padding = 'same', kernel_initializer = 'he_normal')(conv8)
-
-        up9 = Conv2D(64, 2, activation = 'relu', padding = 'same', kernel_initializer = 'he_normal')(UpSampling2D(size = (2,2))(conv8))
-        merge9 = merge([conv1,up9], mode = 'concat', concat_axis = 3)
-        conv9 = Conv2D(64, 3, activation = 'relu', padding = 'same', kernel_initializer = 'he_normal')(merge9)
-        conv9 = Conv2D(64, 3, activation = 'relu', padding = 'same', kernel_initializer = 'he_normal')(conv9)
-        conv9 = Conv2D(2, 3, activation = 'relu', padding = 'same', kernel_initializer = 'he_normal')(conv9)
-        conv10 = Conv2D(1, 1, activation = 'sigmoid')(conv9)
-
-        model = Model(input = inputs, output = conv10)
-        #@TODO add dice coefficient as custom metric
-        model.compile(optimizer = Adam(lr = 1e-4), loss = 'binary_crossentropy', metrics = ['accuracy'])#dice_coef])
-        
-        return model
-    
-    def predict(self, model, test_gen, y_test_fn, view, set_type):
-         
-        for i in range(len(test_gen)):
-            # Temp:
-
-            x_batch,y_batch = test_gen[i]
-            predicted_mask = model.predict_on_batch(x=x_batch) 
-
-            # Save predicted masks
-            self.save_predictions([y_test_fn[i]],
-                                  predicted_mask,
-                                  view=view,
-                                  set_type=set_type)
-            
-    def save_predictions(self, file_names, predictions, view, set_type):
-        # Iterate through filenames and save predictions
-        for idx, fn in enumerate(file_names):
-            pred_fn = fn.replace('mask', 'mask_pred')
-            pred_fp = pred_fn.replace((set_type + '/'), (set_type+'_pred/'))
-            #print('predicted fn : ', pred_fp)
-            self.save_img(predictions[idx], fn = pred_fp)
-            print(pred_fp, ' saved!') 
-
-    def save_img(self, img, fp = '.', fn = 'd_mask.nii'):
-        img_nii = nib.Nifti1Image(img, np.eye(4))
-        img_nii.to_filename(os.path.join(fp, fn))
-        return True
-
-if __name__ == '__main__':
-    myunet = myUnet()
-    myunet.get_unet()
-
-
