+++ utils.py
@@ -1,53 +1,55 @@
-from metrics import dice_coef
-from data_loader import *
+''' Authors : Julien H. & Anish K.
+'''
+
+import os
+from random import shuffle
 import nibabel as nib
+import numpy as np
 
-def computeAverageDice(thresh, pred_dir, gt_dir, n_samples = 500):
+def load_partitioned_data(data_dir, split, y_label='_mask'):
+    ''' 
+    Input -> data_dir : str, split : (float, float, float), y_label : str
+    Load data ito train, val, test set w. specified split.
     '''
-    1) Extract predicted masks from slice_data_side_pred
-    2) Extract masks from slice_data_side
-    3) Compute dice_coef
-    4) Append to list
-    5) return mean
+    total_files = get_file_list_from_dir(data_dir, y_label=y_label)
+    shuffle(total_files) 
+    train_split, val_split, test_split = data_split(total_files, split)
+    x_train = [i[0] for i in train_split]
+    y_train = [i[1] for i in train_split]
+    x_val = [i[0] for i in val_split]
+    y_val = [i[1] for i in val_split]
+    x_test = [i[0] for i in test_split]
+    y_test = [i[1] for i in test_split]
+    return x_train, y_train, x_val, y_val, x_test, y_test
+
+def get_file_list_from_dir(data_dir, y_label='_mask'):
     '''
-    # Retrieve ground truth masks
-    input_mask_pairs = get_file_list_from_dir(gt_dir, y_label='_mask')
-    print(str(len(input_mask_pairs)) + ' in validation set.')
-    # List of dice coefficients
-    dc_list = []
-    idx = 0
-    for input_x, mask_y in input_mask_pairs:
-        # Check if n_samples is reached
-        if n_samples == idx:
-            break
-        # Construct predicted filepath
-        mask_pred_fn = mask_y.replace('_mask', '_mask_pred')
-        mask_pred_fp = mask_pred_fn.replace(gt_dir, pred_dir)
-        if not os.path.isfile(mask_pred_fp):
-            continue
+    Input -> data_dir : str, y_label : str
+    Load filenames from the specified directory.
+    '''
+    total_files = []
+    for _,_,files in os.walk(data_dir):
+        for file in files:
+            if y_label[1:] in file:
+                non_defaced = file.replace(y_label, '')
+                defaced = file
+                total_files.append((data_dir + '/' + non_defaced, data_dir + '/' + defaced))
+        break
+    return total_files
 
-        # Read in files
-        mask_gt = (nib.load(mask_y)).get_data()
-        # If dir. doesn't contain side
-        if 'side' not in pred_dir:
-            mask_pred = ((nib.load(mask_pred_fp)).get_data())[:, :mask_gt.shape[1], :]
-        else:
-            mask_pred = (nib.load(mask_pred_fp)).get_data()
-        # Dice Coefficient computation
-        dc = dice_coef(mask_gt, mask_pred, threshold = thresh)
-        dc_list.append(dc)
-        # Compute dice coef.
-        
-        if idx%100 == 0:
-            print('Avg. Dice Coefficient : ', np.mean(dc_list))
-            #print(dc)
-            print('Example Number', idx)
-        
-        idx+=1
-        
-    return(np.mean(dc_list))
-   
-        
-        
+def data_split(file_list, split):
+    '''
+    Input -> file_list : [str], split : (float, float, float)
+    Split data fn into appropriate data set. 
+    '''	
+    train, val, test = split
+    train_end = train / 100
+    val_end = train_end + (val / 100)
+    test_start = val_end 
+    number_of_files = len(file_list)
 
-avg_dc = computeAverageDice(0.5, './slice_data_side_val_pred', './slice_data_side_val', n_samples = 500)
+    train_split = file_list[:int(train_end * number_of_files)]
+    val_split = file_list[int(train_end * number_of_files):int(number_of_files*val_end)]
+    test_split = file_list[int(number_of_files*test_start):]
+
+    return train_split, val_split, test_split
